NotificationAPIStub deleted and amended into ApiStubs, ApiTestSupport updated to reflect
